+++ .prettierrc.json
@@ -1,5 +1,5 @@
 {
-  "singleQuote": false,
-  "semi": true,
-  "trailingComma": "all"
+"singleQuote": false,
+"semi": true,
+"trailingComma": "all"
 }
