+++ src/rule.js
@@ -1,0 +1,3 @@
+const carNameRule = /^[a-zA-Z0-9]{1,9}$/;
+
+export default carNameRule;
